Casts to remove compiler warnings across architectures

+++ linux.cc
@@ -1225,13 +1225,13 @@ FORK_TEST(Linux, ProcessClocks) {
   // down.
   EXPECT_OK(clock_gettime(child_clock, &ts));
   if (verbose) fprintf(stderr, "[parent: %d] clock_gettime(child=%d->0x%08x) is %ld.%09ld \n",
-                       self, child, child_clock, ts.tv_sec, ts.tv_nsec);
+                       self, child, child_clock, (long)ts.tv_sec, (long)ts.tv_nsec);
 
   child_clock = ((~1) << 3) | 0x0;
   memset(&ts, 0, sizeof(ts));
   EXPECT_OK(clock_gettime(child_clock, &ts));
   if (verbose) fprintf(stderr, "[parent: %d] clock_gettime(init=1->0x%08x) is %ld.%09ld \n",
-                       self, child_clock, ts.tv_sec, ts.tv_nsec);
+                       self, child_clock, (long)ts.tv_sec, (long)ts.tv_nsec);
 
   // Orphan the child.
 }
